思维导图界面优化

+++ android/mindpin/src/com/mindpin/database/FeedDraft.java
@@ -1,8 +1,10 @@
 package com.mindpin.database;
 
 import java.util.ArrayList;
+
+import com.mindpin.Logic.AccountManager;
+import com.mindpin.Logic.Global;
 import android.content.ContentValues;
-import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
@@ -14,9 +16,10 @@ public class FeedDraft {
 	public String select_collection_ids;
 	public boolean send_tsina;
 	public long time;
+	public int user_id;
 
 	public FeedDraft(int id,String title, String content, String image_paths,
-			String select_collection_ids,boolean send_tsina,long time) {
+			String select_collection_ids,boolean send_tsina,long time,int user_id) {
 		this.id = id;
 		this.title = title;
 		this.content = content;
@@ -24,18 +27,19 @@ public class FeedDraft {
 		this.select_collection_ids = select_collection_ids;
 		this.send_tsina = send_tsina;
 		this.time = time;
+		this.user_id = user_id;
 	}
 	
-	public static int get_count(Context context){
-		SQLiteDatabase db = get_read_db(context);
+	public static int get_count(){
+		SQLiteDatabase db = get_read_db();
 		Cursor cursor = db.query(Constants.TABLE_FEED_DRAFTS,new String[]{Constants.KEY_ID}, null, null, null, null, null);
 		int count = cursor.getCount();
 		db.close();
 		return count;
 	}
 	
-	public static FeedDraft find(Context context,int fid){
-		SQLiteDatabase db = get_read_db(context);
+	public static FeedDraft find(int fid){
+		SQLiteDatabase db = get_read_db();
 		Cursor cursor = db.query(Constants.TABLE_FEED_DRAFTS,
 				new String[]{
 								Constants.KEY_ID,
@@ -44,7 +48,8 @@ public class FeedDraft {
 								Constants.TABLE_FEED_DRAFTS__IMAGE_PATHS,
 								Constants.TABLE_FEED_DRAFTS__SELECT_COLLECTION_IDS,
 								Constants.TABLE_FEED_DRAFTS__SEND_TSINA,
-								Constants.TABLE_FEED_DRAFTS__TIME
+								Constants.TABLE_FEED_DRAFTS__TIME,
+								Constants.TABLE_FEED_DRAFTS__USER_ID
 							}, 
 							Constants.KEY_ID + " = "+fid, null, null, null,null);
 		boolean has = cursor.moveToFirst();
@@ -58,15 +63,16 @@ public class FeedDraft {
 			boolean send_tsina = false;
 			if(cursor.getInt(5) == 1) send_tsina = true;			
 			long time = cursor.getLong(6);
-			FeedDraft fh = new FeedDraft(id,title, content, image_paths, select_collection_ids,send_tsina,time);
+			int user_id = cursor.getInt(7);
+			FeedDraft fh = new FeedDraft(id,title, content, image_paths, select_collection_ids,send_tsina,time,user_id);
 			return fh;
 		}else{
 			return null;
 		}
 	}
 	
-	public static ArrayList<FeedDraft> get_feed_drafts(Context context){
-		SQLiteDatabase db = get_read_db(context);
+	public static ArrayList<FeedDraft> get_feed_drafts(){
+		SQLiteDatabase db = get_read_db();
 		Cursor cursor = db.query(Constants.TABLE_FEED_DRAFTS,
 				new String[]{
 								Constants.KEY_ID,
@@ -75,7 +81,8 @@ public class FeedDraft {
 								Constants.TABLE_FEED_DRAFTS__IMAGE_PATHS,
 								Constants.TABLE_FEED_DRAFTS__SELECT_COLLECTION_IDS,
 								Constants.TABLE_FEED_DRAFTS__SEND_TSINA,
-								Constants.TABLE_FEED_DRAFTS__TIME
+								Constants.TABLE_FEED_DRAFTS__TIME,
+								Constants.TABLE_FEED_DRAFTS__USER_ID
 							}, 
 				null, null, null, null,Constants.KEY_ID+ " asc");
 		ArrayList<FeedDraft> fhs = new ArrayList<FeedDraft>();
@@ -88,41 +95,42 @@ public class FeedDraft {
 			boolean send_tsina = false;
 			if(cursor.getInt(5) == 1) send_tsina = true;	
 			long time = cursor.getLong(6);
-			FeedDraft fh = new FeedDraft(id,title, content, image_paths, select_collection_ids,send_tsina,time);
+			int user_id = cursor.getInt(7);
+			FeedDraft fh = new FeedDraft(id,title, content, image_paths, select_collection_ids,send_tsina,time,user_id);
 			fhs.add(fh);
 		}
 		db.close();
 		return fhs;
 	}
 	
-	private static SQLiteDatabase get_write_db(Context context){
-		MindpinDBHelper md = new MindpinDBHelper(context, Constants.DATABASE_NAME,
+	private static SQLiteDatabase get_write_db(){
+		MindpinDBHelper md = new MindpinDBHelper(Global.application_context, Constants.DATABASE_NAME,
 				null, Constants.DATABASE_VERSION);
 		return md.getWritableDatabase();
 	}
 	
-	private static SQLiteDatabase get_read_db(Context context){
-		MindpinDBHelper md = new MindpinDBHelper(context, Constants.DATABASE_NAME,
+	private static SQLiteDatabase get_read_db(){
+		MindpinDBHelper md = new MindpinDBHelper(Global.application_context, Constants.DATABASE_NAME,
 				null, Constants.DATABASE_VERSION);
 		return md.getReadableDatabase();
 	}
 	
-	public static void destroy_all(Context context){
-		SQLiteDatabase db = get_write_db(context);
-		db.execSQL("DELETE FROM "+ Constants.TABLE_FEED_DRAFTS);
+	public static void destroy_all(int user_id){
+		SQLiteDatabase db = get_write_db();
+		db.execSQL("DELETE FROM "+ Constants.TABLE_FEED_DRAFTS+" WHERE"+Constants.TABLE_FEED_DRAFTS__USER_ID+" = ?",new Object[]{user_id});
 		db.close();
 	}
 
-	public static void destroy(Context context,int id) {
-		SQLiteDatabase db = get_write_db(context);
-		db.execSQL("DELETE FROM "+ Constants.TABLE_FEED_DRAFTS +" WHERE "+Constants.KEY_ID+" = ?", new Object[]{id});
+	public static void destroy(int feed_id) {
+		SQLiteDatabase db = get_write_db();
+		db.execSQL("DELETE FROM "+ Constants.TABLE_FEED_DRAFTS +" WHERE "+Constants.KEY_ID+" = ?", new Object[]{feed_id});
 		db.close();
 	}
 
-	public static void update(Context context, int feed_draft_id, String feed_title,
+	public static void update(int feed_draft_id, String feed_title,
 			String feed_content, String images_str,
 			String select_collection_ids_str,boolean send_tsina) {
-		SQLiteDatabase db = get_write_db(context);
+		SQLiteDatabase db = get_write_db();
 		
 		ContentValues values = new ContentValues();
 		values.put(Constants.TABLE_FEED_DRAFTS__TITLE,feed_title);
@@ -137,10 +145,13 @@ public class FeedDraft {
 		db.close();
 	}
 
-	public static void insert(Context context, String title, String content,
+	public static void insert( String title, String content,
 			String images_str, String select_collection_ids_str,boolean send_tsina) {
-		SQLiteDatabase db = get_write_db(context);
-		
+		SQLiteDatabase db = get_write_db();
+		int user_id = AccountManager.current_user_id();
+		if(user_id == 0){
+			return;
+		}
 		ContentValues values = new ContentValues();
 		values.put(Constants.TABLE_FEED_DRAFTS__TITLE,title);
 		values.put(Constants.TABLE_FEED_DRAFTS__CONTENT,content);
@@ -149,6 +160,7 @@ public class FeedDraft {
 		int send_tsina_int = send_tsina?1:0;
 		values.put(Constants.TABLE_FEED_DRAFTS__SEND_TSINA,send_tsina_int);
 		values.put(Constants.TABLE_FEED_DRAFTS__TIME,System.currentTimeMillis());
+		values.put(Constants.TABLE_FEED_DRAFTS__USER_ID,user_id);
 		db.insert(Constants.TABLE_FEED_DRAFTS,null, values);
 		db.close();
 	}
